pyproject.toml
[tool.black]
line-length = 119
skip-string-normalization = true
target-version = ['py39']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | plana\/settings
    | plana\/apps\/associations\/migrations
    | plana\/apps\/commissions\/migrations
    | plana\/apps\/contents\/migrations
    | plana\/apps\/documents\/migrations
    | plana\/apps\/groups\/migrations
    | plana\/apps\/institutions\/migrations
    | plana\/apps\/projects\/migrations
    | plana\/apps\/users\/migrations
    | fabfile
    | _build
    | buck-out
    | build
    | dist
  )/
  | manage.py
  | tox.ini
  | .coveragerc
                     # the root of the project
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.poetry]
name = "plana"
version = "0.9.20230914"
description = ""
authors = ["DIP - Université de Strasbourg <dnum-dip@unistra.fr>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
django = "<5.0"
django-allauth = "<0.55" # higher versions broke django-allauth-cas
# django-allauth-cas = "^1.0.0" # deprecated package for Django 4, need to use a pull request
django-allauth-cas = { url = "https://github.com/thejoeejoee/django-allauth-cas/archive/5db34b546eb32524a3a1a4b90f411e370ac7ad9b.zip" }
django-cleanup = "^8.0.0"
django-cors-headers = "^4.2.0"
django-extensions = "^3.2.3"
django-storages = {extras = ["boto3"], version = "^1.14.0"}
django-summernote = "<0.9"
django-thumbnails = "^0.7.0"
djangorestframework = "^3.14.0"
djangorestframework-camel-case = "^1.4.2"
djangorestframework-simplejwt = "^5.3.0"
dj-rest-auth = "^5.0.0"
drf-spectacular = "^0.26.4"
britney = "^0.5.1"
num2words = "^0.5.12"
openpyxl = "^3.1.2"
psycopg2-binary = "^2.9.7"
pyrage = "^1.1.1"
urllib3 = "<2.0" # needed to avoid dependency problems when updating boto3 and botocore
weasyprint = "^59.0.0"
zxcvbn = "^4.4.28"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
coverage = "^7.3.1"
django-debug-toolbar = "^4.2.0"
pre-commit = "^3.4.0"
pylint = "^2.17.5"
pylint-django = "^2.5.3"
safety = "^2.3.5"
sentry-sdk = "^1.30.0"
tox = "<4.5" # higher versions broke safety

[tool.poetry.group.prod.dependencies]
chaussette-backport = "1.3.1"
sentry-sdk = "^1.30.0"
waitress = "^2.1.2"

[tool.poetry.group.preprod.dependencies]
chaussette-backport = "^1.3.1"
sentry-sdk = "^1.30.0"
waitress = "^2.1.2"

[tool.poetry.group.test.dependencies]
chaussette-backport = "^1.3.1"
sentry-sdk = "^1.30.0"
waitress = "^2.1.2"
